+++ plugins/ibis/superduper_ibis/data_backend.py
@@ -96,10 +96,13 @@ class IbisDataBackend(BaseDataBackend):
         self.overwrite = False
         self._setup(conn)
 
+        self.datatype_presets = {'vector': 'superduper.ext.numpy.encoder.Array'}
+
         if uri.startswith('snowflake://'):
             self.bytes_encoding = 'base64'
-
-        self.datatype_presets = {'vector': 'superduper.ext.numpy.encoder.Array'}
+            self.datatype_presets = {
+                'vector': 'superduper.components.datatype.NativeVector'
+            }
 
     def _setup(self, conn):
         self.dialect = getattr(conn, "name", "base")