+++ plugins/snowflake/plugin_test/test_vector_search.py
@@ -3,7 +3,7 @@ from test.utils.usecase.vector_search import build_vector_index
 
 import pytest
 from superduper import CFG, superduper
-from superduper.components.vector_index import vector
+from superduper.components.datatype import Vector
 
 from superduper_snowflake.vector_search import SnowflakeVectorSearcher
 
@@ -21,8 +21,11 @@ except Exception as e:
 @pytest.mark.skipif(DO_SKIP, reason='Only snowflake deployments relevant.')
 def test_basic_snowflake_search():
     CFG.vector_search_engine = 'snowflake'
+    CFG.force_apply = True
+
     db = superduper()
-    d1 = vector(shape=[300])
+
+    d1 = Vector(shape=[300])
     build_vector_index(db, n=10, list_embeddings=True, vector_datatype=d1, measure='l2')
 
     vector_index = "vector_index"