+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecate vanilla `DataType`
 - Remove `_Encodable` from project
 - Connect to Snowflake using the incluster oauth token
+- Add postprocess in apibase model.
 
 #### New Features & Functionality
 
@@ -37,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the issue where the trigger fails in custom components.
 - Fix serialization between vector-search client and vector-search backend with `to_dict`
 - Fix the bug in the update diff check that replaces uuids
+- Fix snowflake vector search issues.
 
 ## [0.4.0](https://github.com/superduper-io/superduper/compare/0.4.0...0.3.0])    (2024-Nov-02)